requirements.txt
flask
requests
emoji>=1.2.0
flask-cors
grpcio
grpcio-reflection
grpcio-health-checking